code updated

+++ src/main/java/com/simplevat/utils/ExcelUtil.java
@@ -14,18 +14,19 @@ import org.apache.poi.poifs.filesystem.POIFSFileSystem;
 import org.apache.poi.ss.usermodel.DataFormatter;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 @Component
 public class ExcelUtil {
 
+	private static final Logger LOGGER = LoggerFactory.getLogger(ExcelUtil.class);
+
 	public final Integer firstRowIndex = 0;
 
 	public Map<String, List<String>> getDataFromExcel(MultipartFile multipartFile) {
-
-		// File SAMPLE_XLSX_FILE_PATH = new
-		// File("C:\\Users\\Daynil\\Downloads\\SampleData\\SampleData.xlsx");
 
 		if (multipartFile != null) {
 
@@ -78,70 +79,16 @@ public class ExcelUtil {
 					excelDateMap.put(indexHeaderMap.get(key), indexDateMap.get(key));
 				}
 				workbook.close();
-				file.deleteOnExit();
+
+				if (file != null)
+					file.deleteOnExit();
 
 				return excelDateMap;
 			} catch (EncryptedDocumentException | IOException | InvalidFormatException e) {
-				e.printStackTrace();
+				LOGGER.error("Error = ", e);
 			}
 
 		}
 		return new HashMap<String, List<String>>();
-	}
-
-	public List<Map<String, String>> getDataFromExcel1(MultipartFile multipartFile) {
-
-		// File SAMPLE_XLSX_FILE_PATH = new
-		// File("C:\\Users\\Daynil\\Downloads\\SampleData\\SampleData.xlsx");
-
-		if (multipartFile != null) {
-//
-//			File file = null;
-//			try {
-//				file = File.createTempFile(multipartFile.getName(), ".xlsx");
-//				multipartFile.transferTo(file);
-//			} catch (IOException e1) {
-//				e1.printStackTrace();
-//			}
-
-			Map<Integer, String> indexHeaderMap = new HashMap<Integer, String>();
-
-			List<Map<String, String>> list = new ArrayList<Map<String, String>>();
-
-			// Creating a Workbook from an Excel file (.xls or .xlsx)
-			try {
-				Workbook workbook = WorkbookFactory.create(multipartFile.getInputStream());
-
-				// Create a DataFormatter to format and get each cell's value as String
-				DataFormatter dataFormatter = new DataFormatter();
-
-				workbook.forEach(sheet -> {
-					System.out.println("=> " + sheet.getSheetName());
-					Map<String, String> dataeMap = new LinkedHashMap<String, String>();
-					sheet.forEach(row -> {
-						row.forEach(cell -> {
-							String cellValue = dataFormatter.formatCellValue(cell);
-							if (cell.getRow().getRowNum() == firstRowIndex) {
-								System.out.print(cellValue + "\t");
-								indexHeaderMap.put(cell.getColumnIndex(), cellValue);
-							} else {
-
-								dataeMap.put(indexHeaderMap.get(cell.getColumnIndex()), cellValue);
-							}
-						});
-						list.add(dataeMap);
-					});
-				});
-
-				workbook.close();
-//				file.deleteOnExit();
-
-				return list;
-			} catch (EncryptedDocumentException | IOException | InvalidFormatException e) {
-				e.printStackTrace();
-			}
-
-		}
-		return new ArrayList<Map<String, String>>();
 	}
 }
